pyproject.toml
[tool.poetry]
name = "pyxiom"
version = "0.1.0"
description = ""
authors = ["tehforsch <mail@tonipeter.de>"]

[tool.poetry.dependencies]
python = "^3.8"
h5py = "^3.7.0"
astropy = "^5.1"
PyYAML = "^6.0"
numpy = "^1.23.3"
matplotlib = "^3.6.0"
click = "^8.1.3"
seaborn = "^0.12.2"
polars = "^0.18.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150